Update interface font name
Fixes issue when titlebar buttons increased in size on hover event.
The reason of the issue is that Medusa (gauge lib) also loads Roboto
font, but with different name (Roboto Regular). This is rather weird
problem, because it's only reproducible on Windows and only when view
that contains gause is the first view that was loaded to the router
pane.